Refactor `SortConfigTests` and `AbstractSorter` to utilize `SortConfig` builder pattern; migrate tests to reflect configuration changes and improve modularity. Add `DescendingTests` for enhanced coverage.
