+++ Teracura.AbstractSort.Tests/ReturnTypeTests.cs
@@ -20,7 +20,8 @@ public class ReturnTypeTests
     {
         const ReturnType type = ReturnType.List;
         List<string?> list = [null, "", " ", "a", "ba"];
-        var obj = list.SortLength(type);
+        var config = new SortConfig.Builder().Build();
+        var obj = list.SortLength(config);
         list.ShouldBe([null, "", " ", "a", "ba"]);
 
         obj.ShouldBeOfType<List<string?>>();
@@ -31,7 +32,8 @@ public class ReturnTypeTests
     {
         const ReturnType type = ReturnType.Queue;
         List<int> list = [1, 1, 2, 10, 9032, 0, -13];
-        var obj = list.SortLength(type);
+        var config = new SortConfig.Builder().ReturnType(type).Build();
+        var obj = list.SortLength(config);
         list.ShouldBe([0, 1, 1, 2, 10, -13, 9032]);
         obj.ShouldBeOfType<Queue<int>>();
     }
@@ -41,7 +43,8 @@ public class ReturnTypeTests
     {
         const ReturnType type = ReturnType.Stack;
         List<string> list = ["aaa", "b", "cccc"];
-        var obj = list.SortLength(type);
+        var config = new SortConfig.Builder().ReturnType(type).Build();
+        var obj = list.SortLength(config);
         obj.ShouldBeOfType<Stack<string>>();
         list.ShouldBe(["b", "aaa", "cccc"]);
     }
@@ -51,7 +54,8 @@ public class ReturnTypeTests
     {
         const ReturnType type = ReturnType.HashSet;
         List<string> list = ["z", "aaa", "bb", "bb", "z"];
-        var obj = list.SortLength(type);
+        var config = new SortConfig.Builder().ReturnType(type).Build();
+        var obj = list.SortLength(config);
         obj.ShouldBeOfType<HashSet<string>>();
         list.ShouldBe(["z", "z", "bb", "bb", "aaa"]);
         var set = (HashSet<string>)obj!;
